(PUP-4659) Fix environment acceptance tests when master has agent role
Previously, the tests failed when the master had the agent role. This
wasn't noticed originally due to PUP-4111. We also stopped testing the
agent role on the master in the lead up to Puppet 4, so once PUP-4111
was fixed, we didn't see the problem in platform CI. But puppetserver CI
does test that combination and was failing.

This commit changes the following environment scenarios:

bad: Changes the expectation to match the non-existent environmentpath
specified in the main section of puppet.conf

default: Calls the `use_an_environment` method passing in a nil
environment, so that the various puppet commands don't specify
`--environment` on the command line. ALso changes the expectation to
match the default environment `production` in the resulting output, e.g.
Preparing to install into /etc/puppetlabs/code/environments/production/modules

master_environmentpath: The test tried to set an environmentpath in the
master section of puppet.conf, which wasn't different than the default,
and verify that non-master puppet commands, e.g. `puppet apply`
would fail. This commit changes the test to use a non-default
environment path for the master section only, and verify that the
various puppet commands don't "see" that directory.

[skip ci]