(SERVER-1751) Check for env expiration in cached loader list method
Previously, when Puppet::Environments::Cached.list was called to get a
list of environments, no precautions were taken to ensure that the list
of settings returned for Environment objects were freshened per the
environment cache, as they would be if the get() method for an
individual environment were called instead.

In this commit, a call to Cached.list() populates a cache entry for any
environment not represented in the cache, removing any entries from the
cache which have expired.  Along with any environment expirations which
occur, the underlying settings cache for each environment is expired.